backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database (Supabase PostgreSQL)
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM provider (Google Gemini)
google-generativeai==0.8.3

# Document processing (for PDF and Word files)
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.11.0

# Security (optional)
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
